Introduce asyncrhonous API requests
This commit introduces the `aio` package, which is a wrapper over
Kenneth Reitz 'requests' library.

It will be needed to 'watch' Kubernetes API endpoints.

It also comes with command line utility to test connections and
responses. For instance, run:

    python -m kuryr_kubernetes.aio.methods get http://localhost:8080

To see Kubernetes API root resource response.

It may make sense to put it in a separated library or in kuryr-lib, but
for now and for the sake of push other patches quick, it is better to be
included here.

Partial-Implements: blueprint kuryr-k8s-integration
Change-Id: Iea64d7c7941351d6e0511cafde8914b86c457b1b
Authored-By: Antoni Segura Puimedon
Co-Authored-By: Jaume Devesa <devvesa@gmail.com>
Signed-off-by: Jaume Devesa <devvesa@gmail.com>