Introduce the translator layer
This commit introduces the translation layer to call Neutron API, as
well as the config_map that glues the watcher layer to the translation
layer, and it makes it configurable.

It simplifies a lot the watcher layer, making trivial to create new ones
and relies all the business logic to the translation layer. That
introduces the option to some people extend the project by creating
custom translators by implementing a simple interface and using another
config_map.

Partial-Implements: blueprint kuryr-k8s-integration
Change-Id: I1c8bcfe1f37f28e833dd2d4631d011f044fd9c47
Signed-off-by: Jaume Devesa <devvesa@gmail.com>