devstack: make cni paths configurable
With this patch, we'll be able to place our cni executables in:

    /opt/stack/cni/bin

It also allows you to put configuration files for the kuryr cni driver
in:

    /opt/stack/cni/conf

For now we are not setting any non standard conf. It is foreseeable
though that some configuration like which host device to base the
ipvlan/vlan/macvlan devices off on container-in-vm environments could go
here (Although it could also be in a kuryr.conf for the cni driver).

Change-Id: I9abc0878612b97795c68c8d773e7a29e891d1a6f
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>