testing: add offline testing container for CI
This patch adds the necessary code for building a container that can be
used for our CI functional and fullstack testing needs. The only
requirement for building is a working docker engine and the OS to have
access to a packaging mirror (which CI is supposed to have even when
internet connectivity is limited).

In order to build, simply move to 'contrib/testing/container' and run:

    ./build.sh

To try it out locally, just do:

    docker build -t kuryr/testing_container .
    docker run --rm -p 8000:8000 kuryr/testing_container

And from another terminal:

    curl 0:8000

Change-Id: Id94ad2d7567f9e5d60fce25d29c8c3d4a391aabc
Signed-off-by: Antoni Segura Puimedon <antonisp@celebdor.com>