Skip stale K8s events
When a new K8s resource (e.g. pod) is created, a series of events are
triggered. The current flow is as follows:
  1. events A and B happen in a short succession; neither of them
     contain Kuryr-related annotation
  2. Kuryr handles event A, creates necessary Neutron resources and
     adds an annotation, which triggers another event C
  3. Kuryr then handles event B, but since the annotations added in
     previous step are not yet available in event B, Kuryr is not
     aware of already created Neutron resources

This patch introduces a workaround that allows to skip events that
happen in short 'bursts' so handlers will only receive the latest
event. A more robust solution will be added in a later patch.

Change-Id: I794b34a7b9518b8212ff14e61e2bde61f44546f9