Resolve 'resourceVersion' conflicts
It is possible for K8s resource to be updated while it is being
processed by Kuryr handler. As a result K8sClient will fail to
update annotations after the handler finishes its work due to the
'resourceVersion' conflict.

This patch updates K8sClient to allow annotation updates with
conflicting 'resourceVersion' if the updated annotations do not
conflict with the current state (i.e. annotations are either new
or have the same value as in the current state of the object).

Change-Id: I6db14169bf8d6a8114a991bca9c220e15fccdce4
Co-Authored-By: Antoni Segura Puimedon <antonisp@celebdor.com>
Closes-Bug: #1662448