Improve pipeline/Async logging
Currently events are grouped by K8s object's UID and that grouping
key is used for logging in Async handler.

This patch changes grouping key to K8s object's 'selfLink' to help
with log readability. It has no impact on functionality since
'selfLink' is as unique as UID as far as Kuryr is concerned.

Change-Id: Ie903f64d181723d5db0f59dcdf282f5a206879a3
Partially-Implements: blueprint kuryr-k8s-integration