Nested vif driver extension to enable ports reuse
In order to speed up containers creation/deletion a new nested
vif pool driver is proposed to ensure subports already created
and attached to the VM trunk port can be reused in the future.

Note this remove the neutron.create_port and the
neutron.attach_subport (if a port is reused) from the container
creation process.

As measured in the performance evaluation performed in [0], just
the API time for create_port is, on average, around 2 seconds.
On top of that, as the attached subports are already in ACTIVE
status, there is no need for waiting for the status to become
ACTIVE, which is where most of the time is spend now.

[0] https://blog.russellbryant.net/2016/12/19/comparing-openstack-neutron-ml2ovs-and-ovn-control-plane/

Partially Implements blueprint ports-pool

Change-Id: Ibae054ef38d7aa7ea3b2829642f9544d61e76798