doc: Add reminder to have https in api_root
api_root defaults to http://localhost:8080. In the section about https
configuration we give directions to set the client side ssl certificate
and key, but do not mention api_root, which could lead forgetful people
to mistakes if they don't have the url properly set with https and the
corresponding encrypted port.

Change-Id: I261984ebaf7a91875030b68234594ccffc482c3e
Related-Bug: #1670346