devstack: Add configuration for kubelet probes
Most pods for real workloads come with a good amount of readiness and
liveness probes that are performed via either http or tcp. The problem
that we have in devstack baremetal is that the kubelet sits on the host
networking and has no way to reach the pods that are on the overlay.

This patch adds a configuration option that if enabled, makes devstack
create an port for the kubelet to use and binds it in ovs.

Closes-bug: #1693378
Partially Implements: devstack-support-api-accessing-pods
Signed-off-by: Antoni Segura Puimedon <antonisp@celebdor.com>
Change-Id: I8929ca43d162d63efc5b5e5fdf601cd336dd5bc6