Delete or recover precreated ports upon controller restart
Upon kuryr-controller restarted the information about the pre-created
ports is lost and therefore the ports remain created on the neutron
side. In order to avoid having ports leftovers, this patch ensures
they are cleaned up for the generic neutron case and they are
recovered and returned to the corresponding pools for the nested one.

Partially Implements blueprint ports-pool

Change-Id: Ic1f18b483130b753b160cf38843a610537b18466