Clarify _is_pending function use
VIFHandler._is_pending's name is currently misleading, since the
function only returns when 1) the node is in PENDING state; AND 2) the
node has a node assigned.
This commit renames the function to better match implementation, adds a
docstring and a unit-test, that tests for a pod pending without a node
assigned.

Change-Id: I880a579354f1f9bf939ca22fbdcd33d0fc442fcd