Copy kubeconfig to tempest dir
In order to allow the tempest user from the upstream gate to load the
kubeconfig needed for the pod tests this script checks if there's such
home directory and if so it copies and mods the .kube/config file.

Otherwise, the tests fail as the .kube/config is written as stack user
in the gate at /opt/stack/new, which is the STACK_USER home.

Change-Id: Ibac0737c557aa1b6eef55d5681fd4c601cec4f09
Signed-off-by: Daniel Mellado <dmellado@redhat.com>