devstack: fix ovs_bind device to be on the pod net
Commit If5a49bc2010ea59965c5ee9570fb5d131cdf4089 introduced a new
network deployment where the private, pod and service nets are split
out. As a result, having the host with a port in the private subnet
didn't help anymore for kubelet probes.

There are two solutions to this:

a) Add a route to the pod subnet
b) Move the ovs bound host device to the pod subnet.

In this patch I opted for (b)

Change-Id: I1e895b6f64a0929421953cf071b217743784f29a
Signed-off-by: Antoni Segura Puimedon <antonisp@celebdor.com>