Add methods to populate/free subport pools
This patch adds the needed methods to be able to populate/free
specific pools for the nested-vlan pool driver. By using this
methods X subports can be created and attached to a given set
of pools (in this case of trunk ports), so that pools can be
pre-populated with the required amount of port in advance,
without the need of allocating any pod. Additionaly, it also
allows to remove all unused subports from the given pool(s).

Change-Id: I1f06427f095429f1149ac2a7cb9dd5341a84674f