Avoid port update neutron call during pods boot up
This patches removes the need to update the port name when it is
associated to a pod. It still allows it by enabling the port_debug
option at kuryr.conf. By default is set to False, which means it
will not set any name for the ports being used by the pods, nor
to the pods available in the port pools. This makes it similar to
VMs booted with Nova, where unless a predefined port is used, the
port will be created and used without setting any name.

Thanks to disabling the port_debug, a call to Neutron to update the
port name is saved, with the consequent speed up in containers boot
up time, as well as minimization of congestion on the Neutron side
in case of many containers being created at the same time.

Partially Implements blueprint port-creation-control-plane-perf
Change-Id: I674fd948569558d3cc8f13f920c1793e056dfd92