Recover precreated ports at NeutronVIFPool driver
This patch allows to recover the precreated ports when using
the NeutronVIFPool driver. Unlike the nested case, the ports
are not assigned to any node/trunk, therefore this patch gets
the available nodes from kubernetes API and evenly distributes
the existing ports among the pools associated to each node.

Implements blueprint bm-recover-ports

Change-Id: I1ded49573d63aa2c5072ca8636f6d196d8922771