Fix kubelet retries issues
When an error happens during kubelet's ADD request kubelet will trigger
a retry. If aforementioned failure happened after the host-side veth
interface was moved into the host netns, all further kubelet retries of
the request will fail with conflict, as it will be impossible to move
another iface of that name into the host netns.

This commit aims to fix the issue by checking for existence of
conflicting interface in the host netns and removing it if needed.

Change-Id: I1596585c8b076d09a7f8c854bb524c2374d804e8
Closes-Bug: 1730644