rm tools/tox_install.sh
This file is completely unused and the new PTI docs job check whether it
exists - and if it does, use it and execute it.

so, now this file is used - and fails to setup.

Remove it to get everything working again.

Change-Id: I863b32e5bccc03cd2ef91f2590b59e24543ba353
(cherry picked from commit e450cfa3bc542fe148cadb72422522b994bf9b9d)