Dockerfiles: COPY code after installing packages
This change moves COPY directive in kuryr-controller.Dockerfile to be
executed after all the required packages are installed through yum. With
it a layer with all the packages installed can be reused when rebuilding
the image with only Kuryr code changes. This makes development easier
and faster as `yum install` is the longest process during building the
docker image.

Change-Id: I94a7401e393d9c8e51ab50967de792aeae15f811