Delete pods from registry in CNI daemon
Initial implementation of kuryr-daemon was only adding new pods to the
in-memory registry and not deleting them. While the amount of data is
tiny, it will accumulate over time and in practice be a memory leak. We
should aim to prevent that, especially taking into account most recent
efforts to base health checks on memory usage.

This commit makes kuryr-daemon react to pod deletions with removals of
pods from the in-memory registry.

Change-Id: I45babc325b7a727d6a973717f6653a40503ad991
Closes-Bug: 1750394