devstack-heat: Make parameters overridable
Up until now, if you wanted to change the defaults in the parameters,
you had to go and edit the hot/parameters.yml file. This was
inconvenient because it would then show up in git diff and prevent git
review from normal operation. This patch makes it so that now you can
pass the parameters just like if it was the openstack-cli. For example:

    ./devstack-heat stack \
        --parameter \
        "public_net=fbf9bcc6-cbaa-4c24-8d56-8010915a6494" \
        --parameter \
        "image=CentOS-7-x86_64-GenericCloud-1801-01"

Change-Id: Icf4fb124d03f56c8cf383971c5161520a1734d94
Signed-off-by: Antoni Segura Puimedon <antonisp@celebdor.com>