Add HTTPS support to K8s API healthchecks
Our Controller and CNI healthchecks are checking the connection to the
K8s API. This is done using the requests library directly and not
through K8sClient wrapper. Because of that we've forgot to add HTTPS
support there (DevStack runs K8s with HTTP).

This commit fix that by making /healthz requests to go through
K8sClient wrapper.

Change-Id: I7d2657dfb123a7641eb838d7ddecde769bbb4318
Closes-Bug: 1759293