Clean up OpenShift DevStack directories on unstack
When doing a sequence of stack-unstack with DevStack configured to
deploy OpenShift it will leave over the data directory of OpenShift.
When stacking again openshift-node will reuse old configuration even if
$DATA_DIR is removed by the user. This causes it to use old certificates
and be unable to connect to the cluster.

This commit fixes it by cleaning up the directories we use:
* Placing all the configs in the $DATA_DIR. Before this commit
  openshift-node data landed in $DEST/openshfit.
* Moving OpenShift binaries directory from $DEST/openshift to
  $DATA_DIR/openshift/bin.
* Making sure $DATA_DIR/openshift is purged on unstacking.

Change-Id: Ia3bbf868b1a77c9afa22bebd779ec5dc646958ce
Closes-Bug: 1759242