Add better octavia coverage at the gate
This patch moves kuryr-kubernetes-tempest-daemon-containerized-lbaasv2
from experimental to check and gate jobs as it has been stable for
enough time.

In addition, to improve the octavia coverage, this patch adds two
new tests to the experimental gate, both of them for testing octavia
with kuryr containerized deployments. One uses octavia in L3 mode and
the other in L2 mode.

Change-Id: I67867a40bf64aa5a08ca22bb980cd9d6c9b03ebf