Service: Decoupling and improvements of LBaaS driver
A Kuryr controller handler associates itself with specific Kubernetes object
kind (e.g: endpoint) and it's responsible for handling the events of this
object.
The Kuryr controller handlers call one or more Kuryr controller drivers to
manage specific aspects of the Kubernetes resource in the OpenStack domain.
This patch makes K8S-Endpoint handler and LBaaS driver decoupled, and also
updates LBaaS driver to be more generic (e.g: support create a pool that
attached to loadbalancer).

After having this change the LBaaS driver could be extended to
support L7 load balancing, means it could be used as the Ingress driver.

Closes-Bug: 1770934

Change-Id: Ifcda04cc0116bf42e79aa4f855dc9df73671b4d9