Provide data timeouts for K8s API LB
Octavia (and LBaaS v2) defaults to 50 seconds as data timeout for
connections. This is obviously not acceptable for our Watcher
connections that hang watching on K8s API for a long time.

As Octavia recently made this configurable, this commit extends that
timeout to an hour, easily enough to run of all Tempest tests.

Change-Id: I3fe24355acf0943d78f2d9760877ad594f36715e
Closes-Bug: 1773340