Really use the test handler
We had a false positive here and it was raising and swallowing some
exceptions. Fixing it exposed the fact that we were not really removing
the resource from the watch so it would loop forever trying to find more
events instead of finishing.

Change-Id: I96bc2d6ac7951c1cbc78c74e810577fc35587d39
Closes-Bug: #1777155
Closes-Bug: #1777632
Signed-off-by: Antoni Segura Puimedon <antonisp@celebdor.com>
(cherry picked from commit 9a8716087dcccc8afa53e9d1c760a2f3033596a8)