Add KURYR_K8S_LBAAS_USE_OCTAVIA=False to LBaaS job
Turns out in plugin.sh we're depending from aformentioned variable being
set. It defaults to True, so when it wasn't set for LBaaS v2 jobs, we've
ended up not doing some of the stuff we should, e.g. creating a K8s
Service for Neutron router IP.

This commit fixes that as well as some minor issues in code creating the
Service for Neutron router IP. Please note that while it might have been
useful to just switch to use `is_service_enabled q-lbaasv2`, we're in
the process of deprecation of LBaaS v2 support, so I decided to just fix
it the easy way.

Change-Id: If962d9df8780ed342e6658d2d9939863ea7e6ed3
Closes-Bug: 1779312