Allow musl based distros to resolve using search
Musl based distros only perform search in case the ndots config
(which defaults to 1) is bigger than the amount of dots in the
query. This means that if you try to lookup a dns name like:
kubernetes.default.svc it will not try to use the search domain
and the communication will fail.

Change-Id: Ic1af364648de3dd401bb5e77800c2c9675f1becf
Signed-off-by: Antoni Segura Puimedon <asegurap@redhat.com>