Make daemon-cont-octavia and ovn gate non-voting
Until [1] has been investigated I'll be making this gate non-voting as
test_vm_service_http fails and thus blocks any patch to merge, even if
we're currently on FF until release.

OVN gate is also giving some trouble related to ovs, so we'll be making
it also non-voting for the time being.

[1] https://bugs.launchpad.net/kuryr-kubernetes/+bug/1787546
[2] https://bugs.launchpad.net/kuryr-kubernetes/+bug/1787563
Change-Id: I18b7267da6d7bae714fa4a5954760dd88e05d54e