Avoid release_vif race when using namespaces and pools
If ports pools and namespace isolation features are enabled together
it may happen that namespace object gets deleted before the vif handler
on_delete function gets executed, leading to a ResourceNotReady
exception raised by the namespace security group driver when trying
to get the kuryrnet crd annotated on the (deleted) namespace object.

This patch ensure if that happens, the vif handler can still complete
the on_deleted function, releasing the vif by using an empty security
group list as the ports are going to be deleted anyway with the rest
of the namespace related OpenStack resources.

Closes-Bug: 1787951
Change-Id: I19a8f89fb13de1b43e2f6178b6b0261f5dde7990