Verify keystone connection using token
This patch simplifies the keystone connectivity check to just
get a token instead of making another call that internally gets
a token too. This ensures that only the public keystone endpoint
needs to be accesible, by contrast with the previous project or
region list that needed access to the admin keystone endpoint,
which is usually not accesible from within OpenStack VMs.

Closes-Bug: 1789632
Change-Id: Id19a7e7e1662c33dfb5df417ac24d22926227737