LBaaS driver: Add provider support
The  OVN provider was added lately to Octavia [1] in addition
to the 'amphora' provider.
In order to create a LB with a specific provider the provider
name should be posted in load balancer creation request,
in case the provider name is not specified, Octavia default
provider will be used.

This patch enhances LBaaS driver to support LB provider functionality.
The functionality of translating K8S-services into Octavia-OVN
load balancers will be based on this change.

[1] https://github.com/openstack/networking-ovn/commit/66502f19f2994512fb89e6e453429e140da5fa4b

Partially Implements: blueprint octavia-ovn-provider

Change-Id: I07532c0fd93d9269f317eace53b4baaa70df9201