Enable `--restart on-failure` for K8s containers
To workaround gate issues with etcd performance that were causing
kubernetes-controller-manager and kubernetes-scheduler pods to die in
the gate, this commit adds `--restart on-failure` to Kubernetes
containers we start in DevStack plugin.

Minor issue here is that watching the logs will not get restarted
because of DevStack's inability to set restart policy for system.d
services.

Change-Id: I73715c14641236dead437357036e227e3e29f041
Closes-Bug: 1798120