Ignore exc on FIP association if state is correct
If we're getting Conflict exception on FIP association we should check
if that FIP isn't already associated correctly and if that's the case
ignore the exception. This commit implements that behavior.

This solves the case when kuryr-controller is restarted after FIP
association but before that fact is saved into LBaaSState annotation.

Change-Id: I1126906516b787a5ce4dc5421f5323ad4f4b331c
Closes-Bug: 1798549