Add missing key check while lbaas member add
Add check for missing key during member addition for lbaas.
Since lbaasv2.0 API does not support UDP and dns seems to attempt
to add it, it fails readyness and liveness checks for
controller pod.

This happens for k8s-service with a mix of UDP and TCP ports.
In future (Stein release) when UDP LB is supported, it should
not be a problem.

Closes-Bug: 1792367

Change-Id: I73d26c43e924438c2eeba0ced185639f28731fea
(cherry picked from commit 5e44d17e87b3a0468e37875730262252544f7ec0)