Lowercase hostname when using it as K8s nodeName
kuryr-daemon is watching K8s API for new pods appearing on it's node. To
identify such pods nodeName property of pod is used. In
non-containerized deployments we're assuming that node's hostname is the
nodeName. Turns out that K8s is lowercasing it before making it the
nodeName.  We should do the same, otherwise kuryr-daemon will never find
it's pods.

Change-Id: I2bc6b78823e888001b93284b6efa962fad42a37b
Closes-Bug: 1806062