Fix labels encode
If more than one label is applied in the 'labelSelector'
the result retrieved from the query corresponds only to the first label
in the lists of labels, since K8S API do not recognize the '&' applied
when the labels are anded in the encode. This commit fix the issue by
replacing with ','.

Change-Id: I5c2154f399667cd3a7acc4925e78ebd32bfb591a
Closes-Bug: 1806911
Partially Implements: blueprint k8s-network-policies